Multivariate model for binary outcomes update
In the previous binary model, the values of beta_2 and beta_3 are  extremely large – so that the proability of the outcome is super high – over 99%. That is why we are not recovering the true value. There is no way to detect the effect.
 
B_0 should be negative, and and B_1, _2, and _3 should be less than 1. And reduce the variance b_0 and b_1. The estimation will go much better.